fix : 체크리스트 페이지 랜더링 시 post
